project/requirements.txt
asgiref==3.5.1
boto3==1.24.3
botocore==1.27.3
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect2==0.6.0
dj-database-url==0.5.0
Django==4.0.4
django-filter==21.1
django-storages==1.12.3
easy-thumbnails==2.8.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
lxml==4.9.0
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
razorpay==1.3.0
reportlab==3.6.10
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
twilio==7.9.0
tzdata==2022.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.2.0
